pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ocr-provider"
version = "0.1.0"
description = "Standalone OCR service for document/image ingest"
requires-python = ">=3.11"
dependencies = [
    "easyocr>=1.7.2",
    "fastapi>=0.115.0",
    "Pillow>=10.4.0",
    "pytesseract>=0.3.13",
    "PyMuPDF>=1.24.10",
    "rapidocr-onnxruntime>=1.4.4",
    "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
benchmark = [
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "reportlab>=4.4.10",
]

[tool.hatch.build.targets.wheel]
packages = ["provider"]
